<gear_upgrade_extension> gear registry bug 1034967 #noupstream
https://bugzilla.redhat.com/show_bug.cgi?id=1034967

The logic was constructing the new gear registry by extracting the UUID
from the gear DNS entry, which is incorrect for 1.1-era scaled app
secondary gears. Now we actually get the UUID directly.

Since this is specific to OSE upgrade and won't run again for Online,
there is no real point in upstreaming the change. Perhaps a little PSA
would be in order, though.