<broker> RFE 1057153: enable customizing default app templates
Enterprise bug 1057153
https://bugzilla.redhat.com/show_bug.cgi?id=1057153

origin-server #cherrypick
commit f0c39e6c016376da182f9279f637013bf819db1b
Author: Luke Meyer <lmeyer@redhat.com>
Date:   Tue Feb 4 18:31:29 2014 -0500

https://trello.com/c/EoO1OFal/200-2-enterprise-admin-can-change-default-app-git-templates
This enables the administrator to configure URLs that should be used as
the default application templates for framework cartridges. This is done
in broker configuration.

Conflicts:
	broker/conf/broker.conf
	broker/config/environments/development.rb
	broker/config/environments/production.rb
	broker/config/environments/test.rb
	controller/app/models/application.rb

Upstream code in this area of application.rb is nearly unrecognizable
compared to OSE at this time. The patch was modified to significantly.